requirements.txt
about-time==4.2.1
absl-py==2.3.1
alive-progress==3.3.0
altair==5.5.0
astunparse==1.6.3
autograd==1.8.0
blinker==1.9.0
cachetools==6.2.2
captum==0.8.0
click==8.3.1
cma==3.2.2
cupy-cuda11x==13.5.1
Deprecated==1.2.18
dill==0.4.0
et_xmlfile==2.0.0
fastrlock==0.8.3
flatbuffers==25.2.10
fpdf2==2.8.5
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
graphemeu==0.7.2
grpcio==1.74.0
h5py==3.14.0
ipython-genutils==0.2.0
jupyter-highlight-selected-word==0.2.0
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_nbextensions_configurator==0.6.4
jupyterlab==4.4.5
jupyterlab_execute_time==3.2.0
keras==3.11.1
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.44.0
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
mealpy==3.0.3
ml_dtypes==0.5.3
munkres==1.1.4
namex==0.1.0
notebook==7.4.5
numpy==1.26.0
openpyxl==3.1.5
opfunu==1.0.4
opt_einsum==3.4.0
optree==0.17.0
protobuf==5.29.5
pyarrow==21.0.0
pydeck==0.9.1
pymoo==0.6.1.1
rich==14.1.0
smmap==5.0.2
streamlit==1.51.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.1.2
zstandard==0.23.0
